The subscriber only works for Programmer objects - so we'll need more event listeners in the future for other classes.
I have a different idea. Imagine we could link via annotations.

That last part is an expression, from Symfony's expression language. You and I are going to build the system that makes this work, so I'm going to assume that we'll pass a variable called object to the expression language that is this Programmer object being serialized.

Of course, this won't work yet - in fact it'll totally bomb if you try it.