Every annotation - like Table or Entity from Doctrine - has a class behind it. That means we need a Link class.
To hook this annotation into the annotations system, we need a few annotations: the first being, um, well, @Annotation. The second is @Target, which will be 'CLASS'. This means that this annotation is expected to live above class declarations

Inside the Link class, we need to add a public property for each option that can be passed to the annotation, like route and params.

The first property becomes the default property, which is why we don't need to have name = 'self' when using it.

The name and route options are required