Time to evaluate the expression.
To use the expression language, we just need to create an ExpressionLanguage object. We could register this as a service, but I'll take a shortcut and instantiate a new expression language right inside the constructor.

For each param, we'll replace it with $ this->expressionLanguage->evaluate(). Pass it $ param - that's the expression. Next, since the expressions are expecting a variable called object, pass an array as the second argument with an object key set to $ object

All tests passing now.